pyproject.toml
[project]
name = "owlclaw"
version = "0.1.0"
description = "AI news digest pipeline: RSS fetch → LLM curation → Obsidian + Slack"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "google-api-python-client>=2.120",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4",
    "pytest>=8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "UP", "I"]
ignore = []
